requirements.txt
asgiref==3.3.1
Django==3.1.4
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-etc==1.2.0
django-hitcount==1.3.3
django-js-asset==1.2.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
